requirements.txt
Flask==0.11.1
gnupg==2.0.2
python-swiftclient==3.0.0
python-keystoneclient==3.1.0
